JWT authenticated

+++ backend/controllers/goalController.js
@@ -1,12 +1,13 @@
 const asyncHandler = require("express-async-handler");
 
 const Goal = require("../models/goalModel");
+const User = require("../models/userModel");
 
 // @desc Get Goals
 // @route GET /api/goals
 // access Private
 const getGoals = asyncHandler(async (req, res) => {
-  const goals = await Goal.find();
+  const goals = await Goal.find({ user: req.user.id });
   res.status(200).json(goals);
 });
 
@@ -20,6 +21,7 @@ const setGoals = asyncHandler(async (req, res) => {
   }
   const goal = await Goal.create({
     text: req.body.text,
+    user: req.user.id,
   });
   res.status(200).json(goal);
 });
@@ -29,10 +31,25 @@ const setGoals = asyncHandler(async (req, res) => {
 // access Private
 const updateGoals = asyncHandler(async (req, res) => {
   const goal = await Goal.findById(req.params.id);
+
   if (!goal) {
     res.status(400);
     throw new Error("Goal doesn't exist");
   }
+
+  const user = await User.findById(req.user.id);
+  //check for user
+  if (!user) {
+    res.status(401);
+    throw new Error("USer not found");
+  }
+
+  //checking if the goal user is the authenticated user
+  if (goal.user.toString() != user.id) {
+    res.status(401);
+    throw new Error("User not authorized!");
+  }
+
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
     new: true,
   });
@@ -48,6 +65,19 @@ const deleteGoals = asyncHandler(async (req, res) => {
     res.status(400);
     throw new Error("Goal doesn't exist");
   }
+  const user = await User.findById(req.user.id);
+  //check for user
+  if (!user) {
+    res.status(401);
+    throw new Error("USer not found");
+  }
+
+  //checking if the goal user is the authenticated user
+  if (goal.user.toString() != user.id) {
+    res.status(401);
+    throw new Error("User not authorized!");
+  }
+
   await Goal.findByIdAndRemove(req.params.id);
   res.status(200).json({ message: `deleted goal with id ${req.params.id}` });
 });
